+++ src/react/ScopeProvider.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { ScopeContext } from "./contexts/ScopeContext";
-import { MoleculeScope } from "./scope";
+import { MoleculeScope } from "../vanilla/scope";
 import { MoleculeScopeOptions, useScopes } from "./useScopes";
 
 export type ProviderProps<T> = {
