+++ src/react/contexts/ScopeContext.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import type { ScopeTuple } from "../types";
+import type { ScopeTuple } from "../vanilla/types";
 
 export const ScopeContext = React.createContext<ScopeTuple<unknown>[]>([]);
 ScopeContext.displayName = "JotaiMoleculeScopeContext";
