+++ src/react/contexts/StoreContext.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { createStore } from "../store";
+import { createStore } from "../vanilla/store";
 
 export const StoreContext = React.createContext(createStore());
 StoreContext.displayName = "JotaiMoleculeStoreContext";
