+++ src/vanilla/TupleAndReferences.tsx
@@ -1,0 +1,7 @@
+import { AnyScopeTuple } from "./types";
+
+
+export type TupleAndReferences = {
+  references: Set<Symbol>;
+  tuple: AnyScopeTuple;
+};
