+++ src/vanilla/index.ts
@@ -1,0 +1,5 @@
+import { createStoreForExport } from "jotai/core/store";
+
+export { molecule, Molecule } from "./molecule";
+export { createScope, MoleculeScope } from "./scope";
+export * from "./types"
