+++ src/vanilla/molecule.ts
