+++ src/vanilla/scope.ts
