+++ src/vanilla/weakCache.ts
